Separate classes in single-class file #14
AppSettings.POINT_RADIUS
Coord.render()
Various imports cleanups
